refactor: create providers/gigachat and move backend-specific to one place
