+++ gpt2giga/api/litellm/models.py
@@ -2,8 +2,8 @@ from typing import Optional
 
 from fastapi import APIRouter, Request
 
-from gpt2giga.app_state import get_gigachat_client
 from gpt2giga.common.exceptions import exceptions_handler
+from gpt2giga.providers.gigachat.client import get_gigachat_client
 
 router = APIRouter(tags=["LiteLLM"])
