+++ gpt2giga/api/openai/files.py
@@ -4,7 +4,7 @@ from typing import Optional
 
 from fastapi import APIRouter, HTTPException, Query, Request, Response
 
-from gpt2giga.app_state import get_file_store, get_gigachat_client
+from gpt2giga.app_state import get_file_store
 from gpt2giga.api.openai.helpers import (
     _load_batch_output_content,
     _paginate_items,
@@ -14,6 +14,7 @@ from gpt2giga.api.openai.openapi import files_openapi_extra
 from gpt2giga.common.exceptions import exceptions_handler
 from gpt2giga.common.request_form import read_request_multipart
 from gpt2giga.protocol.batches import map_openai_file_purpose
+from gpt2giga.providers.gigachat.client import get_gigachat_client
 
 router = APIRouter(tags=["OpenAI"])
