+++ gpt2giga/common/streaming.py
@@ -8,9 +8,10 @@ from gigachat import GigaChat
 from gigachat.models import Chat, ChatV2
 from starlette.requests import Request
 
-from gpt2giga.app_state import get_gigachat_client, get_response_store
+from gpt2giga.app_state import get_response_store
 from gpt2giga.common.tools import map_tool_name_from_gigachat
 from gpt2giga.core.logging.setup import rquid_context
+from gpt2giga.providers.gigachat.client import get_gigachat_client
 
 
 async def stream_chat_completion_generator(
