+++ gpt2giga/providers/gigachat/__init__.py
@@ -1,0 +1,19 @@
+"""GigaChat provider integration helpers."""
+
+from gpt2giga.providers.gigachat.auth import (
+    create_gigachat_client_for_request,
+    pass_token_to_gigachat,
+)
+from gpt2giga.providers.gigachat.client import (
+    close_app_gigachat_client,
+    create_app_gigachat_client,
+    get_gigachat_client,
+)
+
+__all__ = [
+    "close_app_gigachat_client",
+    "create_app_gigachat_client",
+    "create_gigachat_client_for_request",
+    "get_gigachat_client",
+    "pass_token_to_gigachat",
+]
